% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_kmeans.R
\name{compute_centroids_c}
\alias{compute_centroids_c}
\title{Aggregate categorical data within clusters}
\usage{
compute_centroids_c(data = dat, cluster = NULL, ambiguous_cutoff = 0.5)
}
\arguments{
\item{data}{A numeric matrix of data where cells correspond to rows and genes correspond to columns}

\item{ambiguous_cutoff}{Proportion of the most represented category in a cluster. Clusters falling below this cutoff will be considered ambiguous.}
}
\value{
A \code{list} that includes the following elements:
\describe{
  \item{cluster}{A \code{vector} of length \code{nrow(data)} containing cluster assignments.}
  \item{d}{A \code{vector} of length \code{nrow(data)} containing density values.}
}
}
\description{
Aggregate categorical data within clusters
}
\author{
Kevin Brulois
}
