% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_kmeans.R
\name{compute_centroids_n}
\alias{compute_centroids_n}
\title{Aggregate numeric data within clusters}
\usage{
compute_centroids_n(data = dat, cluster = NULL, centering_function = median)
}
\arguments{
\item{data}{A numeric matrix of data where cells correspond to rows and genes correspond to columns}

\item{centering_function}{A function used to aggregate data for a given gene within a given cluster}
}
\value{
A \code{list} that includes the following elements:
\describe{
  \item{cluster}{A \code{vector} of length \code{nrow(data)} containing cluster assignments.}
  \item{d}{A \code{vector} of length \code{nrow(data)} containing density values.}
}
}
\description{
Aggregate numeric data within clusters
}
\author{
Kevin Brulois
}
