% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_kmeans.R
\name{db_kmeans}
\alias{db_kmeans}
\title{Density-based k-means clustering}
\usage{
db_kmeans(data = dat, clusters = 400, d = NULL, iter_max = 100, label = "")
}
\arguments{
\item{data}{A numeric matrix of data where cells correspond to rows and genes correspond to columns}

\item{clusters}{The number of clusters}

\item{d}{A numeric vector of density values corresponding to rows of data.}

\item{iter_max}{The maximum number of iterations allowed.}

\item{label}{(Optional) A label to be prepended to cluster names and used in messages}
}
\value{
A \code{list} that includes the following elements:
\describe{
  \item{cluster}{A \code{vector} of length \code{nrow(data)} containing cluster assignments.}
  \item{d}{A \code{vector} of length \code{nrow(data)} containing density values.}
}
}
\description{
This algorithm performs k-means clustering and sub-clustering such that the size of each cluster is proportional to the local density. This is done iteratively and the initial size of the kmeans clusters is updated until the desired number of total clusters is reached.
}
\author{
Kevin Brulois
}
