% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADEutilities.R
\name{dbSpadeDownsample}
\alias{dbSpadeDownsample}
\title{A flexible version of SPADE's density-based downsampling function}
\usage{
dbSpadeDownsample(
  data,
  d = NULL,
  target_number = 400,
  target_pctile = NULL,
  target_percent = NULL
)
}
\arguments{
\item{data}{A numeric matrix of data where cells correspond to rows and genes correspond to columns}
}
\value{
A \code{vector} of indices corresponding to rows of \code{data}
}
\description{
This is a version of SPADE's density-based downsampling function, adapted for use with standard R matrices instead of FCS files.
}
\author{
Kevin Brulois
}
