% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchPublicData.R
\name{extractButcherBEC}
\alias{extractButcherBEC}
\title{Download GSE140348 and format as SingleCellExperiment}
\usage{
extractButcherBEC(local.dir = tempdir())
}
\arguments{
\item{local.dir}{A directory where data should be downloaded. Default: tempdir().}
}
\value{
A \code{SingleCellExperiment}
}
\description{
Download GSE140348 and format as SingleCellExperiment
}
\author{
Kevin Brulois
}
