% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchPublicData.R
\name{extractPLN1}
\alias{extractPLN1}
\title{Download a subset of GSE140348 and format as SingleCellExperiment}
\usage{
extractPLN1(local.dir = tempdir())
}
\arguments{
\item{local.dir}{A directory where data should be downloaded. Default: tempdir().}
}
\value{
A \code{SingleCellExperiment}
}
\description{
Download a subset of GSE140348 and format as SingleCellExperiment
}
\author{
Kevin Brulois
}
