% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalDownsample.R
\name{graphicalDownsample}
\alias{graphicalDownsample}
\title{Graphically-based Downsampling}
\usage{
graphicalDownsample(
  data = dat,
  target_number = 400,
  iter_max = 100,
  label = ""
)
}
\arguments{
\item{data}{A numeric matrix of data where cells correspond to rows and genes correspond to columns}

\item{target_number}{The number of cells to downsample to}

\item{iter_max}{The maximum number of iterations allowed.}

\item{label}{(Optional) A label to be used in messages}
}
\value{
A \code{list} that includes the following elements:
\describe{
  \item{indices}{A \code{vector} of indices of \code{data}}
  \item{lattice}{A \code{data.frame} of lattice coordinates used to select data points}
}
}
\description{
This method starts by constructing a lattice encompassing the data and identifies the points in the data that are closest to one of the lattice points. 
This is done iteratively and the distance between lattice points is adjusted until the number of closest data points is approximately equal to a user-specified target_number.
}
\author{
Kevin Brulois
}
