% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_htmlPlot.R
\name{html_3dPlot}
\alias{html_3dPlot}
\title{Generate 3D scatterplot html widgets with color changer}
\usage{
html_3dPlot(
  coordinates = NULL,
  color = NULL,
  discrete_colors_default = readRDS(system.file("data/jmp_discrete_colors.rds", package
    = "Dufy")),
  discrete_colors_custom = NULL,
  continuous_color_scale = viridis::viridis(100, option = "inferno"),
  convert_integer_to_string = TRUE,
  texts = NULL,
  arrows = NULL,
  wrap = NULL,
  include_all_var = TRUE,
  save_widget = TRUE,
  capture_2D = TRUE,
  out_dir = ".",
  file_name = "plot"
)
}
\arguments{
\item{coordinates}{A m x 3 matrix of numeric xyz coordinates}

\item{color}{A m x n data.frame of data to plot as color. Categorical data (factors and character vectors) will be mapped to a discrete color scheme. Numerical data will be mapped to a continuous color scale.}

\item{discrete_colors_default}{A character vector of hexidecimal color codes to use as discrete colors.}

\item{discrete_colors_custom}{Optional. A custom set of descrete color schemes specified as a list of named vectors of hexidecimal color codes. The list names must match a colname of color. The names attribute of each list element must coorespond to each unique element within the matched color column.}

\item{continuous_color_scale}{A character vector of hexidecimal color codes to use as a continuous color scale.}

\item{convert_integer_to_string}{Whether to convert integer columns of color to character. Default: TRUE}

\item{texts}{Optional. A string cooresponding to a colname of color containing a categorical variable. If specified,}

\item{arrows}{Optional. A p x q data.frame containing arrow coordinates. The base of the arrows must be specified as xyz coodinates in columns 1:3. The tip of the arrows must be provided as xyz coordinates in column 4:6. Two additional optional columns can be included: .}

\item{wrap}{Optional. A string cooresponding to a colname of color containing a categorical variable.}

\item{include_all_var}{The number of nearest neighbors used to construct the knn graph}

\item{save_widget}{Whether to save a html widget. Default: TRUE}

\item{capture_2D}{Whether to capture a 2D orthogonal projection of the current 3d view.}

\item{out_dir}{Number of diffusion map components to compute.}

\item{file_name}{A character string passed to the "command" arugment of the system2 function in order to invoke python. Usually "/usr/local/bin/python3" on a Mac.}
}
\value{
A \code{matrix} containing a 2D orthagonal projection of the current view of the data.
}
\description{
This function generates an interactive 3d scatterplot. Includes a color changer and
multi-panel wrapping.
}
\author{
Kevin Brulois
}
