% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToDimReduction.R
\name{mapToDimReduction}
\alias{mapToDimReduction}
\title{Use nearest neighbor relationships to approximate the positions of new data points within pre-computed low-dimensional representations of a reference dataset}
\usage{
mapToDimReduction(
  refData,
  newData,
  int_dat = NULL,
  refReducedDims,
  nn_k = 10,
  min_k = 4,
  nn_span = 0.1
)
}
\arguments{
\item{refData}{A matrix where cells correspond to rows and features correspond to columns.}

\item{newData}{A matrix where cells correspond to rows and features correspond to columns.}

\item{int_dat}{A matrix of batch corrected data. If left unset, batch correction will be performed on refData and newData using the batchelor::fastMNN function.}

\item{refReducedDims}{A list where each element is a dimensionality reduction of refData.}

\item{nn_k}{For a given new data point, the number of nearest neighbors in the reference data to use.}

\item{min_k}{The minimum number of nearest neighbors out of nn_k nearest neighbors from which to select the "best" combination of nearest neighbors. If set, mapping will be performed using the combination of l (where min_k =< l <= nn_k) nearest neighbors in the reference data whose centroid is the closest to its corresponding new data point in batch-corrected space. Recommended for use with nn_k < 12 as the number of combinations (and computation time) increases exponentially with the size of nn_k.}

\item{nn_span_max}{For a given refReducedDim, nn_span is the ratio between the diagnal spanning the nearest neighbors and the diagnal spanning all data points. If set and if nn_span of the reference nearest neighbors for a given new data point is greater than \code{nn_span_max}, then nearest neighbors will be iteratively removed by selecting the most distant neighbor (defined as the nn with the largest distance to the average the other nn) and recomputing nn_span until falls below \code{nn_span_max}.}
}
\value{
A \code{list} of the same length as refReducedDims
}
\description{
Use nearest neighbor relationships to approximate the positions of new data points within pre-computed low-dimensional representations of a reference dataset
}
\author{
Kevin Brulois
}
