% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTuner.R
\name{monocle3_leiden_wrapper}
\alias{monocle3_leiden_wrapper}
\title{A resTuner-compatible wrapper function for leiden clustering using monocle3}
\usage{
monocle3_leiden_wrapper(input_data, resolution, k, ...)
}
\arguments{
\item{input_data}{A matrix where features correspond to rows and cells correspond to columns.}

\item{resolution}{Either a positive integer specifying the exact number of clusters desired or a vector of length two specifying range.}

\item{k}{The number of rounds of clusterings to perform. Default: 5.}

\item{...}{Arguments passed to monocle3:::leiden_clustering}
}
\value{
A vector of length \code{ncol(input_data)} containing cluster assignments.
}
\description{
A resTuner-compatible wrapper function for leiden clustering using monocle3
}
\author{
Kevin Brulois
}
