% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADEutilities.R
\name{optiDensity}
\alias{optiDensity}
\title{Optimized version of SPADE's local density computation}
\usage{
optiDensity(data, selection_method = c("ks.test", "variance"))
}
\arguments{
\item{data}{A numeric matrix of data where cells correspond to rows}
}
\value{
A \code{vector} of density values corresponding to each row of \code{data}.
}
\description{
This function computes local density using the spade::SPADE.density function and a range of values for the kernel_mult and apprx_mult parameters. It then selects and returns the density with the greatest variance.
}
\author{
Kevin Brulois
}
